requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==24.2.0
billiard==4.2.1
Booktype==1.5
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.2
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
kombu==5.4.2
magic-filter==1.0.12
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
prompt_toolkit==3.0.48
propcache==0.2.0
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.1.1
setuptools==75.2.0
simplejson==3.19.3
six==1.16.0
SQLAlchemy==2.0.36
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.16.0
